# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Admin API
  description: Control plane for commercial VPN (AmneziaWG).
  version: 0.1.0-rc.1
paths:
...
  /api/v1/peers:
    get:
      tags:
      - peers
      summary: List Peers
      description: 'List peers (devices). peer_id = device.id, node_id = device.server_id.

        Create peer: use POST /api/v1/devices/issue (with server_id or omit for load-balanced).

        Remove peer: use POST /api/v1/devices/{id}/revoke or /reset.'
      operationId: list_peers_api_v1_peers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: node_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by node (server) id
          title: Node Id
        description: Filter by node (server) id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: active | revoked
          title: Status
        description: active | revoked
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/peers/{peer_id}:
    get:
      tags:
      - peers
      summary: Get Peer
      description: Get one peer (device) by id.
      operationId: get_peer_api_v1_peers__peer_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: peer_id
        in: path
        required: true
        schema:
          type: string
          title: Peer Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerListItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/servers/{server_id}:
    get:
      tags:
      - servers
      summary: Get Server
      operationId: get_server_api_v1_servers__server_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
      - servers
      summary: Delete Server
      description: Remove a server from inventory. Fails with 409 if devices reference
        it.
      operationId: delete_server_api_v1_servers__server_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags:
      - servers
      summary: Update Server
      operationId: update_server_api_v1_servers__server_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/servers/telemetry/summary:
    get:
      tags:
      - servers
      summary: Get Servers Telemetry Summary
      description: 'Return telemetry summary per server (Prometheus: vpn_node_health,
        vpn_node_peers).'
      operationId: get_servers_telemetry_summary_api_v1_servers_telemetry_summary_get
      security:
      - HTTPBearer: []
      parameters:
      - name: region
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Region
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServersTelemetrySummaryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/servers/{server_id}/logs:
    get:
      tags:
      - servers
      summary: Get Server Logs
      description: Recent action log lines for this server (from agent action logs).
        For container stdout use Docker Telemetry.
      operationId: get_server_logs_api_v1_servers__server_id__logs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      - name: tail
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Tail
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerLogsOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/servers/{server_id}/telemetry:
    get:
      tags:
      - servers
      summary: Get Server Telemetry
      description: Peers count, traffic summary from node. Cache first, live fallback.
        Graceful N/A on failure.
      operationId: get_server_telemetry_api_v1_servers__server_id__telemetry_get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerTelemetryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/servers/{server_id}/status:
    get:
      tags:
      - servers
      summary: Get Server Status
      operationId: get_server_status_api_v1_servers__server_id__status_get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerStatusOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/servers/{server_id}/peers:
    post:
      tags:
      - servers
      summary: Create Server Peer
      description: Create a peer on the chosen server; return one-time config_download_url
        and qr_payload.
      operationId: create_server_peer_api_v1_servers__server_id__peers_post
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminIssuePeerRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminIssuePeerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
    get:
      tags:
      - servers
      summary: Get Server Peers
      description: Fetch peers from node with pubkey, handshake, rx/tx, status. Returns
        empty when node unreachable.
      operationId: get_server_peers_api_v1_servers__server_id__peers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerPeersOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PeerListItemOut:
      properties:
        peer_id:
          type: string
          title: Peer Id
        node_id:
          type: string
          title: Node Id
        user_id:
          type: integer
          title: User Id
        subscription_id:
          type: string
          title: Subscription Id
        public_key:
          type: string
          title: Public Key
        client_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Name
        status:
          type: string
          title: Status
        issued_at:
          type: string
          format: date-time
          title: Issued At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
      type: object
      required:
      - peer_id
      - node_id
      - user_id
      - subscription_id
      - public_key
      - client_name
      - status
      - issued_at
      - revoked_at
      title: PeerListItemOut
      description: Single peer in list (device as peer).
    PeerListOut:
      properties:
        peers:
          items:
            $ref: '#/components/schemas/PeerListItemOut'
          type: array
          title: Peers
        total:
          type: integer
          title: Total
      type: object
      required:
      - peers
      - total
      title: PeerListOut
      description: Peers list response.
...
    ServerOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        region:
          type: string
          title: Region
        api_endpoint:
          type: string
          title: Api Endpoint
        vpn_endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Vpn Endpoint
        public_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Public Key
        status:
          type: string
          enum:
          - online
          - offline
          - degraded
          - unknown
          title: Status
        is_active:
          type: boolean
          title: Is Active
        health_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Health Score
        is_draining:
          type: boolean
          title: Is Draining
          default: false
        max_connections:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Connections
        created_at:
          type: string
          format: date-time
          title: Created At
        last_seen_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Seen At
        last_snapshot_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Snapshot At
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        auto_sync_enabled:
          type: boolean
          title: Auto Sync Enabled
          default: false
        auto_sync_interval_sec:
          type: integer
          title: Auto Sync Interval Sec
          default: 60
        ops_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Ops Notes
        ops_notes_updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ops Notes Updated At
        ops_notes_updated_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Ops Notes Updated By
        cert_fingerprint:
          anyOf:
          - type: string
          - type: 'null'
          title: Cert Fingerprint
        cert_expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cert Expires At
      type: object
      required:
      - id
      - name
      - region
      - api_endpoint
      - vpn_endpoint
      - public_key
      - status
      - is_active
      - created_at
      title: ServerOut
...
    ServerUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        region:
          anyOf:
          - type: string
          - type: 'null'
          title: Region
        api_endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Endpoint
        vpn_endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Vpn Endpoint
        public_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Public Key
        ops_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Ops Notes
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        auto_sync_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Auto Sync Enabled
        auto_sync_interval_sec:
          anyOf:
          - type: integer
          - type: 'null'
          title: Auto Sync Interval Sec
      type: object
      title: ServerUpdate
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Bot API key (X-API-Key header).
    AgentToken:
      type: apiKey
      in: header
      name: X-Agent-Token
      description: Node-agent shared token. mTLS required at reverse-proxy.
  responses:
    ErrorResponse:
      description: Unified error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'